+++ src/frontend/src/pages/ChangeRequestsPage/ChangeRequestsTable.tsx
@@ -3,26 +3,21 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { useHistory, useLocation, useRouteMatch } from 'react-router-dom';
 import { DataGrid, GridColDef, GridRow, GridRowProps, GridToolbar } from '@mui/x-data-grid';
 import { routes } from '../../utils/routes';
 import { datePipe, fullNamePipe, wbsPipe } from '../../utils/pipes';
 import { useAllChangeRequests } from '../../hooks/change-requests.hooks';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import ErrorPage from '../ErrorPage';
-import { Add } from '@mui/icons-material';
-import PageTitle from '../../layouts/PageTitle/PageTitle';
-import { useAuth } from '../../hooks/auth.hooks';
 import { useTheme } from '@mui/system';
-import { useEffect, useMemo, useState } from 'react';
-import { ChangeRequest, ChangeRequestType, isGuest, validateWBS, WbsNumber } from 'shared';
+import { useState } from 'react';
+import { ChangeRequest, ChangeRequestType, validateWBS, WbsNumber } from 'shared';
 import { GridColDefStyle } from '../../utils/tables';
-import { NERButton } from '../../components/NERButton';
-import { Link, Tab, Tabs } from '@mui/material';
+import { Link } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
+import ChangeRequestsTitle from './ChangeRequestsTitle';
 
 const ChangeRequestsTable: React.FC = () => {
-  const history = useHistory();
   const { isLoading, isError, data, error } = useAllChangeRequests();
   if (localStorage.getItem('cr-table-row-count') === null) {
     localStorage.setItem('cr-table-row-count', '50');
@@ -36,29 +31,7 @@ const ChangeRequestsTable: React.FC = () => {
     headerAlign: 'center'
   };
 
-  const auth = useAuth();
   const theme = useTheme();
-
-  // Values that go in the URL depending on the tab value
-  const viewUrlValues = useMemo(() => ['overview', 'all'], []);
-
-  const match = useRouteMatch<{ tabValueString: string }>(`${routes.CHANGE_REQUESTS}/:tabValueString`);
-  const tabValueString = match?.params?.tabValueString;
-
-  // Default to the "all change request" tab
-  const initialValue: number = viewUrlValues.indexOf(tabValueString ?? 'all');
-  const [tabIndex, setTabIndex] = useState<number>(initialValue);
-
-  // Change tab when the browser forward/back button is pressed
-  const { pathname } = useLocation();
-  useEffect(() => {
-    const newTabValue: number = viewUrlValues.indexOf(tabValueString ?? 'all');
-    setTabIndex(newTabValue);
-  }, [pathname, setTabIndex, viewUrlValues, tabValueString]);
-
-  const handleTabChange = (_event: React.SyntheticEvent, newValue: number): void => {
-    setTabIndex(newValue);
-  };
 
   if (isLoading || !data) return <LoadingIndicator />;
 
@@ -158,40 +131,7 @@ const ChangeRequestsTable: React.FC = () => {
   ];
   return (
     <div>
-      <div style={{ marginBottom: 15 }}>
-        <PageTitle
-          title={'Change Requests'}
-          previousPages={[]}
-          tabs={
-            <Tabs value={tabIndex} onChange={handleTabChange} variant="standard" aria-label="change-request-tabs">
-              <Tab
-                label="Overview"
-                aria-label="overview"
-                value={0}
-                component={RouterLink}
-                to={`${routes.CHANGE_REQUESTS}/overview`}
-              />
-              <Tab
-                label="All Change Requests"
-                aria-label="all-change-requests"
-                value={1}
-                component={RouterLink}
-                to={`${routes.ALL_CHANGE_REQUESTS}`}
-              />
-            </Tabs>
-          }
-          actionButton={
-            <NERButton
-              variant="contained"
-              disabled={isGuest(auth.user?.role)}
-              startIcon={<Add />}
-              onClick={() => history.push(routes.CHANGE_REQUESTS_NEW)}
-            >
-              New Change Request
-            </NERButton>
-          }
-        />
-      </div>
+      <ChangeRequestsTitle></ChangeRequestsTitle>
       <DataGrid
         autoHeight
         disableSelectionOnClick
